Added some utils to scream_std_meta.hpp (some moved from scream_kokkos_utils.hpp)
